database Article added, update_user, delete_user added and tested

+++ db/db_user.py
@@ -2,6 +2,7 @@ from db.hash import Hash
 from sqlalchemy.orm import Session
 from schemas import UserBase
 from db.models import DbUser
+from fastapi import HTTPException, status
 
 def create_user(db: Session, request: UserBase):
     new_user = DbUser(
@@ -18,10 +19,17 @@ def get_all_users(db: Session):
     return db.query(DbUser).all()
 
 def get_user(db: Session, id: int):
-    return db.query(DbUser).filter(DbUser.id == id).first()
+    user = db.query(DbUser).filter(DbUser.id == id).first()
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, 
+                            detail=f"User with id {id} not found")
+    
 
 def update_user(db: Session, id: int, request: UserBase):
-    user = db.query(DbUser).filter(DbUser.id == id).first()
+    user = db.query(DbUser).filter(DbUser.id == id)
+    if not user.first():
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, 
+                            detail=f"User with id {id} not found")
     user.username = request.username
     user.email = request.email
     user.password = Hash.bcrypt(request.password)
@@ -30,6 +38,9 @@ def update_user(db: Session, id: int, request: UserBase):
 
 def delete_user(db: Session, id: int):
     user = db.query(DbUser).filter(DbUser.id == id).first()
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, 
+                            detail=f"User with id {id} not found")
     db.delete(user)
     db.commit()
     return "deleted"
